move runQuery to drivers/index
